+++ interp/main_test.go
@@ -2,6 +2,7 @@ package interp
 
 import (
 	"bytes"
+	"reflect"
 	"testing"
 )
 
@@ -108,6 +109,7 @@ var containstests = []struct {
 	{[]byte("foo"), "foo", false},
 	{[]byte("$${foo}"), "foo", false},
 	{[]byte("$foo"), "foo", false},
+	{[]byte("${foo-_bar}"), "foo-_bar", true},
 }
 
 func TestContains(t *testing.T) {
@@ -120,8 +122,36 @@ func TestContains(t *testing.T) {
 
 		if tt.expected != actual {
 			t.Errorf(
-				"Contains(%s, %s): expected: %t, actual: %t",
+				`Contains("%s", "%s"): expected: %t, actual: %t`,
 				string(tt.input), tt.name, tt.expected, actual,
+			)
+		}
+	}
+}
+
+var findtests = []struct {
+	input    []byte
+	expected []string
+}{
+	{[]byte(""), []string{}},
+	{[]byte("${}"), []string{}},
+	{[]byte("foo"), []string{}},
+	{[]byte("$foo"), []string{}},
+	{[]byte("${foo}"), []string{"foo"}},
+	{[]byte("${f-o-o}"), []string{"f-o-o"}},
+	{[]byte("${f_o_o}"), []string{"f_o_o"}},
+	{[]byte("${foo}${bar}"), []string{"foo", "bar"}},
+	{[]byte("${foo}${FOO}"), []string{"foo", "FOO"}},
+	{[]byte("_-${foo}.  ${bar} baz"), []string{"foo", "bar"}},
+}
+
+func TestFindPotentialVariables(t *testing.T) {
+	for _, tt := range findtests {
+		actual := FindPotentialVariables(tt.input)
+		if !reflect.DeepEqual(tt.expected, actual) {
+			t.Errorf(
+				`FindPotentialVariables("%s"): expected: %v, actual %v`,
+				string(tt.input), tt.expected, actual,
 			)
 		}
 	}